Final UAF exploit, increased speed
Removed the original heap spray as making the heap deterministic is not required for a use-after-free. In addition, the object spray count was lowered from 5000 to 500.

This is because the freed allocation's address will be stored in the free list, so all one needs to do is allocate objects to take its place; it does not matter what objects come before and/or after ours.